requirements.txt
ansible~=5.2.0
matrix-nio~=0.19.0
signedjson~=1.1.4